refactor: campsite splits so no over-threshold file grows (#3095, epic #1116)
The #3095 fix pushed three files past/further past the 800-line source target, which
the gate's file-size ratchet FAILs. Split by responsibility instead of waiving:

* cqlite-flight/src/bypass.rs 843 -> 440 (+ bypass_tests.rs), the crate's
  established #[path] unit-test split (admission_tests.rs, batch_bytes_tests.rs);
* cqlite-flight/src/producer_stream.rs 805 -> 403 (+ producer_stream_tests.rs);
* row_decoder/block_emit_windowed.rs 1189 -> 972 (< its 1075 pre-change size) by
  moving the streaming-scan SlidingPartitionPolicy into timestamp_policy.rs.

Also documents the two additive-optional oracle fields
(physical_static_only_partitions, ddl) in query-semantics-oracle.json's doc block.
No behavior change: cqlite-core --lib 3363 pass, cqlite-flight --lib 367 pass.
