+++ test-data/query-semantics-oracle.json
@@ -22,6 +22,19 @@
     "`expected_rows.len()`) a direct demonstration of the physical-vs-semantic gap",
     "this oracle exists to close.",
     "",
+    "`physical_static_only_partitions` (additive-optional, default 0; issue #3095)",
+    "is the count of golden partitions holding a `static_block` and NO `row` entry.",
+    "Cassandra returns exactly ONE result row for each such partition, and that row",
+    "has no physical `row` behind it — so it is the ONLY way a correct semantic",
+    "result may legitimately EXCEED `physical_row_count`. Both lanes re-derive the",
+    "count from the golden and assert this declaration, so the allowance can never",
+    "be inflated to mask an over-emitting read path.",
+    "",
+    "`ddl` (additive-optional) is the keyspace-qualified `CREATE TABLE` the FLIGHT",
+    "lane's ticket must carry, required for any table whose shape differs from the",
+    "`(id int, ck int, v text, PRIMARY KEY (id, ck))` shape shared by every table",
+    "in `compaction-tombstone-ttl-parity.cql` (e.g. a `static` column).",
+    "",
     "Consumed by cqlite-core/tests/query_semantics_oracle_parity.rs."
   ],
   "oracle_version": 1,
